requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==1.2.0
bokeh==0.12.4
chardet==2.3.0
cycler==0.10.0
decorator==4.0.11
discord.py==0.16.7
gcloud==0.17.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
ipython-genutils==0.2.0
Jinja2==2.9.5
jsonschema==2.6.0
jupyter-core==4.3.0
jws==0.1.3
MarkupSafe==1.0
matplotlib==2.0.0
multidict==2.1.4
nbformat==4.3.0
numpy==1.12.1
oauth2client==3.0.0
packaging==16.8
pandas==0.19.2
plotly==2.0.8
protobuf==3.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycryptodome==3.4.3
pyparsing==2.2.0
Pyrebase==3.0.27
python-dateutil==2.6.0
python-jwt==2.0.1
pytz==2016.10
PyYAML==3.12
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.10.0
tornado==4.4.2
traitlets==4.3.2
websockets==3.2
